Use nuint for CpuHandle

+++ src/Vortice.Direct3D12/CpuDescriptorHandle.cs
@@ -90,7 +90,7 @@ public partial struct CpuDescriptorHandle : IEquatable<CpuDescriptorHandle>
     public static bool operator !=(CpuDescriptorHandle left, CpuDescriptorHandle right) => (left.ptr != right.ptr);
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public static implicit operator ulong(CpuDescriptorHandle left) => left.ptr;
+    public static implicit operator nuint(CpuDescriptorHandle left) => left.ptr;
 
     /// <inheritdoc/>
     public override int GetHashCode() => ptr.GetHashCode();
